+++ src/TraceEvent/TraceEventStacks.cs
@@ -63,7 +63,7 @@ namespace Microsoft.Diagnostics.Tracing.Stacks
                 m_log = events.Log;
             m_goodTopModuleIndex = ModuleFileIndex.Invalid;
             m_curSample = new StackSourceSample(this);
-            m_curSample.Metric = (float)events.Log.SampleProfileInterval.TotalMilliseconds;
+            m_curSample.Metric = events.Log.SampleProfileInterval.TotalMilliseconds;
             m_events = events;
             m_maxPseudoStack = m_log.CodeAddresses.Count;     // This really is a guess as to how many stacks we need.  
         }
@@ -464,7 +464,7 @@ namespace Microsoft.Diagnostics.Tracing.Stacks
             m_log = log;
             m_goodTopModuleIndex = ModuleFileIndex.Invalid;
             m_curSample = new StackSourceSample(this);
-            m_curSample.Metric = (float)log.SampleProfileInterval.TotalMilliseconds;
+            m_curSample.Metric = log.SampleProfileInterval.TotalMilliseconds;
         }
 
         // Sometimes we just have a code address and thread, but no actual ETW stack.  Create a 'one element'
